feat: add `session-token` attribute to `user-joined` event
Add the `session-token` attribute to `user-joined` event when available.
The payload is an array of strings (session tokens).

Here, it's most likely going to be a 1-sized array for the foreseeable
future since only the first "join" event for an user is handled (i.e.:
no multi-session handling).

Keeping it an array should make things future-proof, though, for when
multi-session scenarios are handled.